Build left panel
I used a display: flex on the body to have my 2 panels side by side.
Then I set the widths of the panels to a certain proportion of the
viewport using a custom property. I used a background-image but I'm not
sure it's the best way because when resizing the image doesn't behave
well.